+++ packages/cli-v3/src/workers/prod/worker-facade.ts
@@ -9,6 +9,10 @@ import {
   HandleErrorFunction,
   DurableClock,
   clock,
+  getEnvVar,
+  logLevels,
+  LogLevel,
+  ZodSchemaParsedError,
 } from "@trigger.dev/core/v3";
 import "source-map-support/register.js";
 
@@ -47,10 +51,18 @@ clock.setGlobalClock(durableClock);
 const tracer = new TriggerTracer({ tracer: otelTracer, logger: otelLogger });
 const consoleInterceptor = new ConsoleInterceptor(otelLogger);
 
+const triggerLogLevel = getEnvVar("TRIGGER_LOG_LEVEL");
+
+const configLogLevel = triggerLogLevel
+  ? triggerLogLevel
+  : importedConfig
+  ? importedConfig.logLevel
+  : __PROJECT_CONFIG__.logLevel;
+
 const otelTaskLogger = new OtelTaskLogger({
   logger: otelLogger,
   tracer: tracer,
-  level: "info",
+  level: logLevels.includes(configLogLevel as any) ? (configLogLevel as LogLevel) : "log",
 });
 
 logger.setGlobalTaskLogger(otelTaskLogger);
@@ -208,8 +220,14 @@ const prodRuntimeManager = new ProdRuntimeManager(zodIpc, {
 
 runtime.setGlobalRuntimeManager(prodRuntimeManager);
 
-zodIpc.send("TASKS_READY", { tasks: getTaskMetadata() }).catch((err) => {
-  console.error("Failed to send TASKS_READY message", err);
+const TASK_METADATA = getTaskMetadata();
+
+zodIpc.send("TASKS_READY", { tasks: TASK_METADATA }).catch((err) => {
+  if (err instanceof ZodSchemaParsedError) {
+    zodIpc.send("TASKS_FAILED_TO_PARSE", { zodIssues: err.error.issues, tasks: TASK_METADATA });
+  } else {
+    console.error("Failed to send TASKS_READY message", err);
+  }
 });
 
 process.title = "trigger-prod-worker";