+++ packages/database/prisma/migrations/20240329142454_add_concurrency_limit_columns/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Organization" ADD COLUMN     "maximumConcurrencyLimit" INTEGER NOT NULL DEFAULT 10;
+
+-- AlterTable
+ALTER TABLE "RuntimeEnvironment" ADD COLUMN     "maximumConcurrencyLimit" INTEGER NOT NULL DEFAULT 10;