+++ packages/integration-kit/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @trigger.dev/integration-kit
 
+## 3.0.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [c702d6a9c]
+- Updated dependencies [b271742dc]
+- Updated dependencies [9af2570da]
+  - @trigger.dev/core@3.0.0-beta.3
+
+## 3.0.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [e3cf456c6]
+  - @trigger.dev/core@3.0.0-beta.2
+
 ## 3.0.0-beta.1
 
 ### Patch Changes