+++ packages/sveltekit/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @trigger.dev/sveltekit
 
+## 3.0.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [b271742dc]
+  - @trigger.dev/sdk@3.0.0-beta.3
+
+## 3.0.0-beta.2
+
+### Patch Changes
+
+- @trigger.dev/sdk@3.0.0-beta.2
+
 ## 3.0.0-beta.1
 
 ### Patch Changes